% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_years_ahead.R
\name{add_years_ahead}
\alias{add_years_ahead}
\title{Add years_ahead column}
\usage{
add_years_ahead(divs)
}
\arguments{
\item{divs}{a \code{data.table} with columns for yield_mean and comb_growth}
}
\description{
Simple helper function to add a years_ahead column using a combined growth metric
}
