% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anaylze_portfolio.R
\name{anaylze_portfolio}
\alias{anaylze_portfolio}
\title{General function for getting performance information from transactions}
\usage{
anaylze_portfolio(transactions, rounding = 4, updates = NULL)
}
\description{
Summarize all of my transactions to easily see both retrospective and prospective performance metrics. Depends on a specfic set of column names so I'm not going to spend a bunch of time documenting this just yet.
}
