% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ann_growth.R
\name{calc_ann_growth}
\alias{calc_ann_growth}
\title{Calculate the annualized growth}
\usage{
calc_ann_growth(end, start, period = 10)
}
\arguments{
\item{end}{ending amount}

\item{start}{starting amount}

\item{period}{time between start and end}
}
\description{
Calculate the annualized growth
}
