% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_income_growth.R
\name{calc_income_growth}
\alias{calc_income_growth}
\title{Calculate annual income growth}
\usage{
calc_income_growth(dat, period = 10)
}
\arguments{
\item{dat}{a list of \code{data.table} objects in the format returned by
\code{get_income_stream}}

\item{period}{period of growth}
}
\description{
Calculate annual income growth
}
