% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_data.R
\name{gather_data}
\alias{gather_data}
\title{Get Historic Price Data & Summaries}
\usage{
gather_data(syms, show_progress = TRUE, sleep_time = 0.25, ...)
}
\arguments{
\item{syms}{Character vector with symbols to pull data on}

\item{show_progress}{Show a progress bar if TRUE else no console output}

\item{sleep_time}{tiingo doesn't appreciate requests coming in extremely
quickly...}

\item{...}{Additional arguments to be passed into \code{riingo::riingo_prices}}
}
\description{
Pull the historic price information for a set of symbols and show many useful
summaries dealing with dividend growth, yield, RSI, and price.
}
