% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_div_data.R
\name{get_div_data}
\alias{get_div_data}
\title{Get basic dividend data}
\usage{
get_div_data(sym)
}
\arguments{
\item{sym}{character vector of length 1 with a single security's symbol}
}
\description{
Make a call using \code{riingo} to get dividend data on a monthly time scale
}
