% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first.R
\name{get_first}
\alias{get_first}
\title{First \code{n} elements}
\usage{
get_first(x, n = 1)
}
\arguments{
\item{x}{vector to get the first \code{n} elements of}

\item{n}{number of elements to return}
}
\description{
Simple helper function to get the first \code{n} elements of a vector
}
