% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_income_stream.R
\name{get_income_stream}
\alias{get_income_stream}
\title{Get historical income information from a security}
\usage{
get_income_stream(sym, orig_date = today() - years(10), orig_amount = 1000)
}
\arguments{
\item{sym}{a character vector with ticker symbols}

\item{orig_date}{starting date of the investment}

\item{orig_amount}{starting investment amount}
}
\value{
a list where each element is a \code{data.table} containing income information
}
\description{
Pull the income information (including dividends) from a security for a given
period of time and a given starting investment amount. The returned
information will give information about the income stream from reinvestment
separately.
}
