% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last.R
\name{get_last}
\alias{get_last}
\title{Last \code{n} elements}
\usage{
get_last(x, n = 1)
}
\arguments{
\item{x}{vector to get the last \code{n} elements of}

\item{n}{number of elements to return}
}
\description{
Simple helper function to get the last \code{n} elements of a vector
}
