% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_income.R
\name{plot_income}
\alias{plot_income}
\title{Generate an income plot for a set of securities}
\usage{
plot_income(
  dat,
  yield = FALSE,
  cost = 1000,
  annual = FALSE,
  main = ggplot2::element_blank()
)
}
\arguments{
\item{dat}{a list of \code{data.tables} in the format that is returned by \code{get_income_stream}}

\item{yield}{show yield on cost instead of absolute amounts}

\item{cost}{original cost of the investments}

\item{annual}{annualize numbers}

\item{main}{title of the plot}
}
\value{
a \code{ggplot2} object with the plot for the securities
}
\description{
Generate a plot that shows the income broken out by security and reinvestment for a given set of securities
}
