% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_income_forecast.R
\name{plot_income_forecast}
\alias{plot_income_forecast}
\title{Income Forecast over time}
\usage{
plot_income_forecast(
  data,
  date_field = "Date",
  forecast_field = "Forward_div",
  y = NULL,
  add_cagr = TRUE,
  n_years = 1
)
}
\arguments{
\item{data}{a \code{data.table} with a field for date and forward income}

\item{date_field}{a character with the date field}

\item{forecast_field}{a character with the income forecast field}

\item{y}{year of interest (NULL defaults to current year)}

\item{add_cagr}{if TRUE adds a smoothed income over time line and prints the
CAGR from the first time period to the last in the upper left.}

\item{n_years}{How many years to include in the plot}
}
\description{
Create a forward income forcast plot over time. Optionally add an exponential
function connecting the first and last time points to calculate CAGR.
}
