% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_div_yield.R
\name{populate_div_yield}
\alias{populate_div_yield}
\title{Populate dividend yield using closest dividend}
\usage{
populate_div_yield(d)
}
\arguments{
\item{d}{a \code{data.table} that includes the columns that come from \code{pull_dat}
(divCash, date, splitFactor, close)}
}
\value{
a \code{data.table} with the same information as \code{d} but with a column for
yield added (and sorted by date if not already)
}
\description{
This function will add a \code{yield} column to a data.table which will be
populated with the dividend yield. This yield uses the daily closing price
and the dividend closest in time to that date (so likely an overestimate for
about a month prior to any historic dividend raise).
}
