% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_dat.R
\name{pull_dat}
\alias{pull_dat}
\title{Pull historic data & convert to data table}
\usage{
pull_dat(sym, start_date = "1970-01-01", ...)
}
\arguments{
\item{sym}{Character vector of length 1 with a security's ticker}

\item{start_date}{Starting date for which prices should be pulled}

\item{...}{any additional arguments to be passed into \code{riingo::riingo_prices}}
}
\description{
Pull the historic price data of a security and set the resulting data to a
\code{data.table} for future processing
}
