% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{Some functions insist on using \code{cat} to print messages to the console. This
will make them quiet by sinking to a temp file}
\usage{
quiet(x)
}
\arguments{
\item{x}{expression to make quiet}
}
\description{
Some functions insist on using \code{cat} to print messages to the console. This
will make them quiet by sinking to a temp file
}
