% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_quote.R
\name{recent_quote}
\alias{recent_quote}
\title{Simple function for getting the last price of a security through \code{riingo}}
\usage{
recent_quote(sym)
}
\arguments{
\item{sym}{character vector of securities to get the quotes of}
}
\description{
Simple function for getting the last price of a security through \code{riingo}
}
